log greeting
